requirements/base.txt
Django==5.2
django-jazzmin==3.0.1
pillow==11.1.0
python-decouple==3.8
django-ckeditor==6.7.2
django-modeltranslation==0.19.13